# openapi.yaml
openapi: 3.0.0
info:
  title: Snake Rivals Arena API
  version: 1.0.0
  description: API for the Snake Rivals Arena game backend.
servers:
  - url: http://localhost:3000
    description: Local development server

components:
  schemas:
    Player:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        highScore:
          type: integer
        gamesPlayed:
          type: integer
        lastPlayed:
          type: string
          format: date-time
      required:
        - id
        - username
        - score
        - highScore
        - gamesPlayed

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        player:
          $ref: '#/components/schemas/Player'
        score:
          type: integer
        mode:
          $ref: '#/components/schemas/GameMode'
        timestamp:
          type: string
          format: date-time
      required:
        - rank
        - player
        - score
        - mode
        - timestamp

    GameMode:
      type: string
      enum:
        - pass-through
        - walls

    LiveGame:
      type: object
      properties:
        id:
          type: string
        player:
          $ref: '#/components/schemas/Player'
        mode:
          $ref: '#/components/schemas/GameMode'
        currentScore:
          type: integer
        startedAt:
          type: string
          format: date-time
        spectators:
          type: integer
      required:
        - id
        - player
        - mode
        - currentScore
        - startedAt
        - spectators

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/Player'
        token:
          type: string
      required:
        - user
        - token

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /auth/login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/signup:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      summary: Logout the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out

  /auth/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '401':
          description: Unauthorized

  /leaderboard:
    get:
      summary: Get top scores
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of scores to return
      responses:
        '200':
          description: List of top scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'

    post:
      summary: Submit a score
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
                mode:
                  $ref: '#/components/schemas/GameMode'
              required:
                - score
                - mode
      responses:
        '200':
          description: Score submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  newRank:
                    type: integer
        '401':
          description: Unauthorized

  /spectate/live:
    get:
      summary: Get list of live games
      responses:
        '200':
          description: List of live games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LiveGame'

  /spectate/live/{gameId}:
    get:
      summary: Watch a specific game
      parameters:
        - in: path
          name: gameId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveGame'
        '404':
          description: Game not found